reorg and run
