+++ crypto/openstark/src/proof_params.rs
@@ -41,7 +41,7 @@ impl ProofParams {
     pub fn suggested(domain_size_log: usize) -> Self {
         let num_threes = (domain_size_log - 8) / 3;
         let mut fri_layout = vec![3; num_threes];
-        if num_threes * 3 == (domain_size_log - 8) {
+        if num_threes * 3 != (domain_size_log - 8) {
             fri_layout.push(domain_size_log - (8 + num_threes * 3));
         }
