--- crypto/openstark/src/vfd_matter/mod.rs
@@ -1,1 +1,0 @@
-pub mod vdf;
